+++ src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! Developed at the [Department of Materials Physics](https://physics.anu.edu.au/research/mp/), Australian National University, Canberra, Australia.
 //!
-//! **zarrs is experimental and in limited production use. Use at your own risk!**
+//! **zarrs is experimental and in limited production use. Use at your own risk! Correctness issues with past versions are [detailed here](#correctness-issues-with-past-versions).**
 //!
 //! A changelog can be found [here](https://github.com/LDeakin/zarrs/blob/main/CHANGELOG.md).
 //!
@@ -67,6 +67,21 @@
 //! ## Zarrs Ecosystem
 //! - [zarrs-ffi](https://github.com/LDeakin/zarrs-ffi): A subset of zarrs exposed as a C API.
 //! - [zarrs_tools](https://github.com/LDeakin/zarrs_tools): Various tools for creating and manipulating Zarr v3 data.
+//!
+//! ## Correctness Issues with Past Versions
+//! - Prior to zarrs [v0.11.5](https://github.com/LDeakin/zarrs/releases/tag/v0.11.5), arrays that used the `crc32c` codec have invalid chunk checksums
+//!   - Such arrays will fail to be read by other zarr implementations that validate checksums
+//!   - These arrays can be read by zarrs if the [validate checksums](crate::config::Config#validate-checksums) global configuration option is disabled
+//! - From zarrs [v0.11.2](https://github.com/LDeakin/zarrs/releases/tag/v0.11.2)-[v0.11.3](https://github.com/LDeakin/zarrs/releases/tag/v0.11.3), the codec configuration of the `crc32c` codec or `bytes` codec (with unspecified endianness) does not conform to the zarr specification
+//!   - Such arrays will likely fail to be read by other zarr implementations
+//!   - Zarrs still supports reading these arrays, but this may become an error in a future release
+//!   - Fixing these arrays only requires a simple metadata correction, e.g.
+//!     - `sed -i -E "s/(^([ tab]+)\"(crc32c|bytes)\"(,?)$)/\2{ \"name\": \"\3\" }\4/" zarr.json`
+//!
+//! [zarrs_tools](https://github.com/LDeakin/zarrs_tools) v0.2.3+ can fix arrays with the above correctness issues with `zarrs_reencode`. Example:
+//! ```bash
+//! zarrs_reencode --ignore-checksums -p 4 array.zarr array_fixed.zarr
+//! ```
 //!
 //! ## Licence
 //! zarrs is licensed under either of