+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Add a global configuration `config::Config` accessible via `config::{get_config,get_config_mut}`
    - Currently it exposes a single configuration option: `validate_checksums` (default: `true`)
+ - Document correctness issues with past versions and how to correct errant arrays in crate root
 
 ## [0.11.5] - 2024-02-05